# openapi.yaml
openapi: 3.0.0
info:
  title: API de Cooperativas
  description: API para gerenciamento de cooperativas de reciclagem
  version: 1.0.0
  contact:
    name: João Victor Marcondes
    email: joao@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Servidor de desenvolvimento
  - url: https://api.cooperativas.com
    description: Servidor de produção
...
paths:
...
  /login:
    post:
      summary: Verificar credenciais
      description: Autentica um usuário e retorna seu cargo
      tags:
        - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
                - senha
              properties:
                user:
                  type: string
                  description: Nome de usuário
                  example: vitoria
                senha:
                  type: string
                  description: Senha do usuário
                  example: "123456"
      responses:
        200:
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuario:
                    type: string
                    example: vitoria
                  cargo:
                    type: string
                    enum: [adm, tesoureira]
                    example: tesoureira
                  cooperativa:
                    type: string
                    example: santa maria
                  cooperativa_id:
                    type: integer
                    example: 1
                  autenticado:
                    type: boolean
                    example: true
        400:
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /consultar:
    post:
      summary: Consultar dados
      description: Consulta dados das tabelas baseado no cargo do usuário
      tags:
        - Consultas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
                - senha
              properties:
                user:
                  type: string
                  example: vitoria
                senha:
                  type: string
                  example: "123456"
      responses:
        200:
          description: Dados consultados com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ConsultaResponse'
                  - type: object
                    properties:
                      usuario:
                        $ref: '#/components/schemas/UsuarioInfo'
        400:
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Erro interno no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /consultar/filtrado:
    post:
      summary: Consultar dados com filtros
      description: Consulta dados de uma tabela específica com filtros avançados
      tags:
        - Consultas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
                - senha
                - tabela
              properties:
                user:
                  type: string
                  example: vitoria
                senha:
                  type: string
                  example: "123456"
                tabela:
                  type: string
                  enum: [recebimento, triagem, prensa, bazar]
                  example: recebimento
                filtros:
                  $ref: '#/components/schemas/Filtros'
      responses:
        200:
          description: Dados consultados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  tabela:
                    type: string
                    example: recebimento
                  filtros_aplicados:
                    $ref: '#/components/schemas/Filtros'
                  total:
                    type: integer
                    example: 10
                  dados:
                    type: array
                    items:
                      type: object
        400:
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recebimento:
    post:
      summary: Registrar recebimento
      description: Registra um novo recebimento de materiais
      tags:
        - Recebimento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - procedencia
                - placa_caminhao
                - peso_total
                - material_tipo
                - recebido_por
                - cooperativa
              properties:
                procedencia:
                  type: string
                  enum: [coleta, mercado]
                  example: coleta
                placa_caminhao:
                  type: string
                  example: ABC-1234
                peso_total:
                  type: number
                  example: 150.5
                material_tipo:
                  type: string
                  example: papel
                recebido_por:
                  type: string
                  example: João
                cooperativa:
                  type: string
                  enum: [santa maria, coopersel]
                  example: santa maria
      responses:
        200:
          description: Dados enviados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Dados enviados com sucesso!
                  horario_envio:
                    type: string
                    format: date-time
                    example: 26/01/2026 14:30:00
        400:
          description: Erro na requisição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /triagem:
    post:
      summary: Registrar triagem
      description: Registra uma nova triagem de materiais
      tags:
        - Triagem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mesa_id
                - material_tipo
                - qntd_bags
                - peso_rejeito
                - cooperativa
              properties:
                mesa_id:
                  type: integer
                  example: 1
                material_tipo:
                  type: string
                  example: plastico
                qntd_bags:
                  type: integer
                  example: 5
                peso_rejeito:
                  type: number
                  example: 10.5
                cooperativa:
                  type: string
                  enum: [santa maria, coopersel]
                  example: santa maria
      responses:
        200:
          description: Triagem registrada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Triagem registrada com sucesso!
        400:
          description: Erro na requisição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /excluir_recebimento:
    delete:
      summary: Excluir registro de recebimento
      description: Remove um registro da tabela de recebimento por ID e cooperativa
      tags:
        - Recebimento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - cooperativa
              properties:
                id:
                  type: integer
                  example: 12
                cooperativa:
                  type: string
                  enum: [santa maria, coopersel]
                  example: santa maria
      responses:
        200:
          description: Registro excluído com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Registro excluído com sucesso da tabela recebimento!
        400:
          description: Dados insuficientes ou inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
        details:
          type: string
          description: Detalhes do erro (opcional)
      example:
        error: Credenciais inválidas
    UsuarioInfo:
      type: object
      properties:
        nome:
          type: string
          example: vitoria
        cargo:
          type: string
          enum: [adm, tesoureira]
          example: tesoureira
...
    ConsultaResponse:
      type: object
      properties:
        message:
          type: string
          example: Dados de todas as cooperativas
        cooperativas:
          type: string
          example: Todas
        cooperativa_id:
          type: integer
          example: 1
        recebimento:
          type: array
          items:
            type: object
        triagem:
          type: array
          items:
            type: object
        prensa:
          type: array
          items:
            type: object
        bazar:
          type: array
          items:
            $ref: '#/components/schemas/Bazar'
        cooperados:
          type: array
          items:
            $ref: '#/components/schemas/Cooperado'
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Bearer {token}"
security:
  - BearerAuth: []
...